Moved EDIFACT cartridge into its own module; bumped EDIFACT XSD major's version so that it's not confused with the obsolete UNEDIFACT XSD; fixed compilation failures caused by the structural changes in the DFDL cartridge; EDIFACT XSD is now inheriting the attributes of the EDI XSD
