--- cartridge/src/main/resources/META-INF/xsd/smooks/edi-2.0.xsd-smooks.xml
@@ -1,74 +1,0 @@
-<?xml version="1.0"?>
-<smooks-resource-list xmlns="https://www.smooks.org/xsd/smooks-1.2.xsd"
-                      default-selector-namespace="https://www.smooks.org/xsd/smooks/edi-2.0.xsd">
-
-    <resource-config selector="parser">
-        <resource>org.smooks.cdr.extension.NewResourceConfig</resource>
-    </resource-config>
-
-    <resource-config selector="parser">
-        <resource>org.smooks.cdr.extension.SetOnResourceConfig</resource>
-        <param name="setOn">resourceType</param>
-        <param name="value">edi-parser</param>
-    </resource-config>
-    <resource-config selector="parser">
-        <resource>org.smooks.cdr.extension.SetOnResourceConfig</resource>
-        <param name="setOn">selector</param>
-        <param name="value">org.xml.sax.driver</param>
-    </resource-config>
-    <resource-config selector="parser">
-        <resource>org.smooks.cdr.extension.MapToResourceConfigFromAttribute</resource>
-        <param name="attribute">indent</param>
-    </resource-config>
-
-    <resource-config selector="unparser">
-        <resource>org.smooks.cdr.extension.NewResourceConfig</resource>
-    </resource-config>
-    <resource-config selector="unparser">
-        <resource>org.smooks.cdr.extension.SetOnResourceConfig</resource>
-        <param name="setOn">resourceType</param>
-        <param name="value">edi-unparser</param>
-    </resource-config>
-    <resource-config selector="unparser">
-        <resource>org.smooks.cdr.extension.MapToResourceConfigFromAttribute</resource>
-        <param name="attribute">unparseOnElement</param>
-        <param name="mapTo">selector</param>
-    </resource-config>
-
-    <resource-config selector="parser,unparser">
-        <resource>org.smooks.cdr.extension.MapToResourceConfigFromAttribute</resource>
-        <param name="attribute">schemaURI</param>
-    </resource-config>
-    <resource-config selector="parser,unparser">
-        <resource>org.smooks.cdr.extension.MapToResourceConfigFromAttribute</resource>
-        <param name="attribute">validateDFDLSchemas</param>
-    </resource-config>
-    <resource-config selector="parser,unparser">
-        <resource>org.smooks.cdr.extension.MapToResourceConfigFromAttribute</resource>
-        <param name="attribute">segmentTerminator</param>
-    </resource-config>
-    <resource-config selector="parser,unparser">
-        <resource>org.smooks.cdr.extension.MapToResourceConfigFromAttribute</resource>
-        <param name="attribute">dataElementSeparator</param>
-    </resource-config>
-    <resource-config selector="parser,unparser">
-        <resource>org.smooks.cdr.extension.MapToResourceConfigFromAttribute</resource>
-        <param name="attribute">compositeDataElementSeparator</param>
-    </resource-config>
-    <resource-config selector="parser,unparser">
-        <resource>org.smooks.cdr.extension.MapToResourceConfigFromAttribute</resource>
-        <param name="attribute">escapeCharacter</param>
-    </resource-config>
-    <resource-config selector="parser,unparser">
-        <resource>org.smooks.cdr.extension.MapToResourceConfigFromAttribute</resource>
-        <param name="attribute">repetitionSeparator</param>
-    </resource-config>
-    <resource-config selector="parser,unparser">
-        <resource>org.smooks.cdr.extension.MapToResourceConfigFromAttribute</resource>
-        <param name="attribute">decimalSign</param>
-    </resource-config>
-    <resource-config selector="parser,unparser">
-        <resource>org.smooks.cdr.extension.MapToResourceConfigFromAttribute</resource>
-        <param name="attribute">triadSeparator</param>
-    </resource-config>
-</smooks-resource-list>
