--- cartridge/src/main/resources/META-INF/xsd/smooks/edifact-1.0.xsd
@@ -1,68 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<xsd:schema xmlns:xsd="http://www.w3.org/2001/XMLSchema"
-            xmlns:smooks="https://www.smooks.org/xsd/smooks-1.2.xsd"
-            xmlns:smooks-edifact="https://www.smooks.org/xsd/smooks/edifact-1.0.xsd"
-            xmlns:smooks-dfdl="https://www.smooks.org/xsd/smooks/dfdl-1.0.xsd"
-            targetNamespace="https://www.smooks.org/xsd/smooks/edifact-1.0.xsd"
-            elementFormDefault="qualified">
-
-    <xsd:import namespace="https://www.smooks.org/xsd/smooks-1.2.xsd"/>
-    <xsd:import namespace="https://www.smooks.org/xsd/smooks/dfdl-1.0.xsd"/>
-
-    <xsd:annotation>
-        <xsd:documentation xml:lang="en">Smooks EDIFACT Resource Configuration</xsd:documentation>
-    </xsd:annotation>
-
-    <xsd:element name="parser" type="smooks-edifact:parser" substitutionGroup="smooks:abstract-reader">
-        <xsd:annotation>
-            <xsd:documentation xml:lang="en">
-                A EDIFACT parser reads an EDIFACT document by using grammar defined in a DFDL schema file and generates the
-                corresponding event stream that Smooks can process.
-            </xsd:documentation>
-        </xsd:annotation>
-    </xsd:element>
-
-    <xsd:element name="unparser" substitutionGroup="smooks:abstract-resource-config"
-                 type="smooks-edifact:unparserVisitor">
-        <xsd:annotation>
-            <xsd:documentation xml:lang="en">
-                A visitor that uses grammar defined in a DFDL schema file to unparse a fragment selected by the
-                "unparseOnElement" attribute.            </xsd:documentation>
-        </xsd:annotation>
-    </xsd:element>
-
-    <xsd:complexType name="parser">
-        <xsd:annotation>
-            <xsd:documentation xml:lang="en">EDIFACT Parser</xsd:documentation>
-        </xsd:annotation>
-        <xsd:complexContent>
-            <xsd:extension base="smooks-dfdl:parser">
-                <xsd:sequence>
-                    <xsd:element name="messages" type="smooks-edifact:messages"/>
-                </xsd:sequence>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="messages">
-        <xsd:sequence>
-            <xsd:element name="message" maxOccurs="unbounded">
-                <xsd:complexType>
-                    <xsd:simpleContent>
-                        <xsd:extension base="xsd:string"/>
-                    </xsd:simpleContent>
-                </xsd:complexType>
-            </xsd:element>
-        </xsd:sequence>
-    </xsd:complexType>
-
-    <xsd:complexType name="unparserVisitor">
-        <xsd:complexContent>
-            <xsd:extension base="smooks-dfdl:unparserVisitor">
-                <xsd:sequence>
-                    <xsd:element name="messages" type="smooks-edifact:messages"/>
-                </xsd:sequence>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-</xsd:schema>
