--- cartridge/src/test/resources/edi/smooks-parser-config.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0"?>
-<smooks-resource-list xmlns="https://www.smooks.org/xsd/smooks-1.2.xsd"
-                      xmlns:edi="https://www.smooks.org/xsd/smooks/edi-2.0.xsd"
-                      xmlns:dfdl="https://www.smooks.org/xsd/smooks/dfdl-1.0.xsd">
-
-    <edi:parser schemaURI="/edi/edi-to-xml-mapping.dfdl.xsd" dataElementSeparator="*" segmentTerminator="%NL;"
-                compositeDataElementSeparator="^">
-    </edi:parser>
-
-</smooks-resource-list>
