--- cartridge/src/test/resources/edifact/smooks-parser-config.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0"?>
-<smooks-resource-list xmlns="https://www.smooks.org/xsd/smooks-1.2.xsd"
-                      xmlns:edifact="https://www.smooks.org/xsd/smooks/edifact-1.0.xsd"
-                      xmlns:dfdl="https://www.smooks.org/xsd/smooks/dfdl-1.0.xsd">
-
-    <edifact:parser schemaURI="/d03b/EDIFACT-Messages.dfdl.xsd">
-        <edifact:messages>
-            <edifact:message>ORDERS</edifact:message>
-            <edifact:message>INVOIC</edifact:message>
-        </edifact:messages>
-    </edifact:parser>
-
-</smooks-resource-list>
