+++ edi-cartridge/src/main/resources/META-INF/xsd/smooks/edi-2.0.xsd
