+++ edi-cartridge/src/test/resources/data/edi-input.txt
