+++ edi-cartridge/src/test/resources/edi-to-xml-mapping.dfdl.xsd
