+++ edi-cartridge/src/test/resources/smooks-unparser-config.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+<smooks-resource-list xmlns="https://www.smooks.org/xsd/smooks-1.2.xsd"
+                      xmlns:edi="https://www.smooks.org/xsd/smooks/edi-2.0.xsd"
+                      xmlns:dfdl="https://www.smooks.org/xsd/smooks/dfdl-1.0.xsd">
+
+    <edi:unparser schemaURI="/edi-to-xml-mapping.dfdl.xsd" unparseOnElement="/test-message" dataElementSeparator="*"
+                  segmentTerminator="%NL;"
+                  compositeDataElementSeparator="^"/>
+
+</smooks-resource-list>
