+++ edifact-cartridge/src/main/resources/META-INF/content-handlers.inf
@@ -1,0 +1,1 @@
+org.smooks.cartridges.dfdl.unparser.DfdlUnparserContentHandlerFactory
