+++ edifact-cartridge/src/test/resources/data/INVOIC_D.03B_Interchange_with_UNA.txt
