+++ edifact-cartridge/src/test/resources/data/ORDERS_D.03B_Interchange.txt
