+++ edifact-cartridge/src/test/resources/smooks-unparser-config.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0"?>
+<smooks-resource-list xmlns="https://www.smooks.org/xsd/smooks-1.2.xsd"
+                      xmlns:edifact="https://www.smooks.org/xsd/smooks/edifact-2.0.xsd"
+                      xmlns:dfdl="https://www.smooks.org/xsd/smooks/dfdl-1.0.xsd"
+                      xmlns:edi="https://www.smooks.org/xsd/smooks/edi-2.0.xsd">
+
+    <edifact:unparser schemaURI="/d03b/EDIFACT-Messages.dfdl.xsd" unparseOnElement="/Interchange">
+        <edifact:messages>
+            <edifact:message>ORDERS</edifact:message>
+            <edifact:message>INVOIC</edifact:message>
+        </edifact:messages>
+    </edifact:unparser>
+
+</smooks-resource-list>
